+++ CHANGELOG.md
@@ -19,6 +19,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added a **Auto** option to the dashboard size formatter - this intelligently selects the most appropriate size to show
 - Now shows episode names when utilizing the scope **Episodes** in the rule previewer
 - Env variable `LOG_RETENTION_DAYS` _(defaults to 30)_ to keep N daily rotated log files
+- Candidates unified view with filters for movie/series
 
 ### Changed
 
@@ -29,6 +30,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Move is only enabled for Series, Seasons, or Movie Versions (not episodes)
 - **Series** cards now show episode count instead of season count (hovering the number will display the season/episode count in a tool tip)
 - Candidates API now supports 'all'
+- Poster sizes for candidates are all now the same and slightly larger
 
 ### Fixed